Separate index from utility JavaScript
