+++ app/javascript/spree/dashboard/controllers/sortable_tree_controller.js
@@ -1,6 +1,6 @@
 import { Controller } from "@hotwired/stimulus"
 import { Sortable } from "sortablejs"
-import { patch } from "../request_utility"
+import { patch } from "../utilities/request_utility"
 
 export default class extends Controller {
   static values = { handle: String }
