+++ app/javascript/spree/dashboard/utilities/bootstrap.js
