+++ app/javascript/spree/dashboard/utilities/request_utility.js
